tests: move GTEST_COUT_COLOR to testutils.h
